`jabber-parse-time': enable fractional seconds.
XEP-0082 states: The fractions of a second are optional and MAY be
ignored if included (although an XMPP protocol extension using the
DateTime profile MAY require the fractions of a second).

Optional, but why not?  Currently, only :rate-time ewoc entries seem
to cause an issue with fractional seconds.  And the :rare-time
timestamps do not use either seconds or minutes resolution.