Re-implement MUC tame presence notifications.
This aims to improve the implementation of c6f42c1 which allows
selective display of MUC presence notification messages.  The older
implementation discarded MUC presence messages when a pattern in
`jabber-muc-decorate-presence-patterns' matched the message, but did
not provide a face for its display.

Rather than dropping messages matching a pattern wit no face, retain
the messages, but hide them from display.  This allows
`jabber-muc-decorate-presence-patterns' to be changed to reveal and/or
rehide these messages.