Suppress :rare-time clutter in low traffic MUCs.
When MUC presence announcements are hidden, lightly trafficed MUC chat
buffers get clogged by superfluous :rate-time reports.

This patch to the :rare-time EWOC printer reaches back across hidden
presence announces, and if it finds that the last visible EWOC is
another :rare-time event, it erases the text for that event.

This seems rather skeevy in that it makes assumptions about how EWOC
nodes are displayed in the chat buffer.  But it does significantly
declutter quiet MUC chat buffers.

With this patch, when a MUC chat buffer quiets down, as single
:rare-time event are left visible at the bottom of the buffer.  This
represents the last :rare-time event generated, usually by a hidden
MUC presence announcement.