tl.el: reorg byline author/action byline logic
the base idea is to separate our logic of author byline, and action
byline. previously author-byline function would indifferently handle
booster if a boost, else author. now it is always an author, and we
implement booster byline separately. this should pave the way for bylines
at top of posts, and should be clearer and simpler also.

- move booster action-byline elt from byline to insert-status (outside byline proper)
- factor out byline-handle/byline-username funs
- factor our byline-booster/byline-booster-str funs

this breaks notifs byline logic. we will fix it subsequently.