fix media-attachment-ids order. FIX #655.
attachments were being posted in the reverse order that the user uploaded
them. we can't just reverse the ids list upon posting though, as that
would happen again when we posted an edit of a post or similar. so we
resolved to try to only ever generate the list in the correct order upon uploading.