totp-auth-get-backend-for-secret: default to an encrypted backend
If a secret has been stored we return the current backend, whether
or not that happens to be encrypted. However even if the user has
configured an unencrypted backend as the default we will not store
a new secret there - we use the first encrypred backend instead.

Bug reported by Morgan Willcock <morgan@ice9.digital>