add suspended check to tl--toot, no display if so. #753.
A user is getting errors in bookmarks view because they bookmarked a toot
by a user that was then suspended. I think the server should not send us
such data, as suspension means a user's toots disappear (though they're
not deleted for 30 days). but whatever, the server is doing so, so we need
to try to handle the situation gracefully.